requirements.txt
transformers
torch
diffusers
Pillow
pyttsx3
